feat(icons): apply neon outline design to app and tray icons
Replace the pale gray (#e8e4f0) tray glyph — which faded into the
taskbar, especially on Windows — with the chosen neon variant: an
outlined coral-gradient heart with solid waveform bars and a faintly
tinted interior. The recording state turns red and gains a bottom-right
dot badge with a punched gap ring, so it stays unambiguous next to the
coral idle state. The app icon adopts the same mark on the dark rounded
square.

Fold the neon palette into scripts/gen-icons.js and drop the proposal
generator and preview assets now that a variant is picked.
