+++ scripts/gen-icons.js
@@ -1,7 +1,8 @@
 #!/usr/bin/env node
-// Generates the app and tray icons (a heart with a waveform — "earheart")
-// as PNGs, with no image dependencies: pixels are rendered with signed
-// distance functions and encoded with a minimal PNG writer on top of zlib.
+// Generates the app and tray icons — an outlined heart with solid waveform
+// bars ("earheart", neon style) — as PNGs, with no image dependencies:
+// pixels are rendered with signed distance functions and encoded with a
+// minimal PNG writer on top of zlib.
 
 const fs = require("node:fs");
 const path = require("node:path");
@@ -84,79 +85,102 @@ const WAVE_BARS = [
   { cx: 0.5, h: 0.09 },
 ];
 
-function inWave(x, y, scale) {
-  return WAVE_BARS.some((bar) =>
-    inBar(x, y, bar.cx * scale, bar.h * scale, 0.075 * scale)
-  );
+function inWave(x, y) {
+  return WAVE_BARS.some((bar) => inBar(x, y, bar.cx, bar.h, 0.075));
 }
 
-/* ---------- renderers ---------- */
+// Recording dot, bottom-right of the glyph.
+function inDot(x, y, r) {
+  const dx = x - 0.58;
+  const dy = y + 0.58;
+  return dx * dx + dy * dy <= r * r;
+}
+
+/* ---------- renderer ---------- */
 
 function mix(a, b, t) {
   return a + (b - a) * t;
 }
 
-// Renders one icon. `style`: "app" (rounded square bg + heart + wave) or
-// "glyph" (transparent bg, solid-color heart + punched-out wave).
-function render(size, style, glyphColor) {
+// opts:
+//   bg    [r,g,b] rounded-square background, or null for transparent (tray)
+//   grad  [[r,g,b] bottom, [r,g,b] top] vertical gradient for heart and wave
+//   dot   true → bottom-right recording dot with a punched gap ring
+function render(size, opts) {
   const rgba = Buffer.alloc(size * size * 4);
   const SS = 4; // supersampling grid (4x4) for clean edges
   for (let py = 0; py < size; py++) {
     for (let px = 0; px < size; px++) {
       let bgHit = 0;
       let heartHit = 0;
+      let innerHit = 0;
       let waveHit = 0;
+      let dotHit = 0;
+      let gapHit = 0;
       for (let sy = 0; sy < SS; sy++) {
         for (let sx = 0; sx < SS; sx++) {
           // Normalized coords: x right, y up, [-1, 1] across the icon.
           const nx = ((px + (sx + 0.5) / SS) / size) * 2 - 1;
           const ny = 1 - ((py + (sy + 0.5) / SS) / size) * 2;
-          if (style === "app" && inRoundedRect(nx, ny, 1, 0.36)) bgHit++;
+          if (opts.bg && inRoundedRect(nx, ny, 1, 0.36)) bgHit++;
           // Heart space: slightly smaller than the canvas, nudged up.
           const hx = nx / 0.62;
           const hy = (ny + 0.06) / 0.62;
           const inH = inHeart(hx, hy);
           if (inH) heartHit++;
-          if (inH && inWave(nx, ny + 0.02, 1)) waveHit++;
+          if (inH && inHeart(hx / 0.78, hy / 0.78)) innerHit++;
+          if (inH && inWave(nx, ny + 0.02)) waveHit++;
+          if (opts.dot) {
+            if (inDot(nx, ny, 0.3)) dotHit++;
+            else if (inDot(nx, ny, 0.44)) gapHit++;
+          }
         }
       }
       const samples = SS * SS;
       const bgA = bgHit / samples;
       const heartA = heartHit / samples;
+      const innerA = innerHit / samples;
       const waveA = waveHit / samples;
+      const dotA = dotHit / samples;
+      const gapA = gapHit / samples;
 
       let r = 0;
       let g = 0;
       let b = 0;
       let a = 0;
-      if (style === "app") {
-        // Dark rounded square.
-        r = 0x1d;
-        g = 0x1a;
-        b = 0x26;
+      if (opts.bg) {
+        r = opts.bg[0];
+        g = opts.bg[1];
+        b = opts.bg[2];
         a = bgA;
       }
-      if (heartA > 0) {
-        // Vertical coral gradient.
-        const t = (1 - py / size) * 0.9;
-        const hr = style === "app" ? mix(0xff, 0xff, t) : glyphColor[0];
-        const hg = style === "app" ? mix(0x47, 0x84, t) : glyphColor[1];
-        const hb = style === "app" ? mix(0x66, 0x9c, t) : glyphColor[2];
-        r = mix(r, hr, heartA);
-        g = mix(g, hg, heartA);
-        b = mix(b, hb, heartA);
-        a = Math.max(a, heartA);
+      // Heart drawn as an outline band; the interior keeps a faint tint so
+      // the silhouette still reads at tray sizes.
+      const t = (1 - py / size) * 0.9;
+      const hr = mix(opts.grad[0][0], opts.grad[1][0], t);
+      const hg = mix(opts.grad[0][1], opts.grad[1][1], t);
+      const hb = mix(opts.grad[0][2], opts.grad[1][2], t);
+      const bodyA = heartA - innerA + innerA * 0.16;
+      if (bodyA > 0) {
+        r = mix(r, hr, bodyA);
+        g = mix(g, hg, bodyA);
+        b = mix(b, hb, bodyA);
+        a = Math.max(a, bodyA);
       }
+      // Solid waveform bars in the same gradient — the hero of the mark.
       if (waveA > 0) {
-        if (style === "app") {
-          // White waveform on the heart.
-          r = mix(r, 0xff, waveA);
-          g = mix(g, 0xff, waveA);
-          b = mix(b, 0xff, waveA);
-        } else {
-          // Punch the waveform out of the glyph so it works at tray sizes.
-          a = a * (1 - waveA);
-        }
+        r = mix(r, hr, waveA);
+        g = mix(g, hg, waveA);
+        b = mix(b, hb, waveA);
+        a = Math.max(a, waveA);
+      }
+      if (opts.dot) {
+        // Gap ring separates the dot from whatever it overlaps.
+        a = a * (1 - gapA) * (1 - dotA);
+        r = mix(r, 0xff, dotA);
+        g = mix(g, 0x20, dotA);
+        b = mix(b, 0x38, dotA);
+        a = Math.max(a, dotA);
       }
       const i = (py * size + px) * 4;
       rgba[i] = Math.round(r);
@@ -170,20 +194,33 @@ function render(size, style, glyphColor) {
 
 /* ---------- output ---------- */
 
+const CORAL = [
+  [0xff, 0x2f, 0x53],
+  [0xff, 0x8c, 0xa3],
+];
+// Recording: red gradient plus the dot badge, unambiguous next to coral.
+const RED = [
+  [0xe6, 0x0e, 0x33],
+  [0xff, 0x5a, 0x63],
+];
+
 const out = path.join(__dirname, "..", "assets");
 fs.mkdirSync(out, { recursive: true });
 
-fs.writeFileSync(path.join(out, "icon.png"), render(512, "app"));
-fs.writeFileSync(path.join(out, "tray.png"), render(32, "glyph", [0xe8, 0xe4, 0xf0]));
+fs.writeFileSync(
+  path.join(out, "icon.png"),
+  render(512, { bg: [0x16, 0x12, 0x1f], grad: CORAL })
+);
+fs.writeFileSync(path.join(out, "tray.png"), render(32, { grad: CORAL }));
 fs.writeFileSync(
   path.join(out, "tray-recording.png"),
-  render(32, "glyph", [0xff, 0x54, 0x70])
+  render(32, { grad: RED, dot: true })
 );
 // macOS tray works best with a 2x variant.
-fs.writeFileSync(path.join(out, "tray@2x.png"), render(64, "glyph", [0xe8, 0xe4, 0xf0]));
+fs.writeFileSync(path.join(out, "tray@2x.png"), render(64, { grad: CORAL }));
 fs.writeFileSync(
   path.join(out, "tray-recording@2x.png"),
-  render(64, "glyph", [0xff, 0x54, 0x70])
+  render(64, { grad: RED, dot: true })
 );
 
 console.log("Icons written to assets/");
